Refactored routes, requests now don't need to pass the user id anymore

+++ src/main/kotlin/com/kuluruvineeth/routes/CommentRoutes.kt
@@ -18,8 +18,7 @@ import io.ktor.server.response.*
 import io.ktor.server.routing.*
 
 fun Route.createComment(
-    commentService: CommentService,
-    userService: UserService
+    commentService: CommentService
 ){
     authenticate {
         post("/api/comment/create"){
@@ -27,37 +26,32 @@ fun Route.createComment(
                 call.respond(HttpStatusCode.BadRequest)
                 return@post
             }
-            ifEmailBelongsToUser(
-                userId = request.userId,
-                validateEmail = userService::doesEmailBelongToUserId
-            ){
-                when(commentService.createComment(request)){
-                    is CommentService.ValidationEvent.ErrorFieldEmpty -> {
-                        call.respond(
-                            HttpStatusCode.OK,
-                            BasicApiResponse(
-                                successful = false,
-                                message = FIELDS_BLANK
-                            )
+            when(commentService.createComment(request,call.userId)){
+                is CommentService.ValidationEvent.ErrorFieldEmpty -> {
+                    call.respond(
+                        HttpStatusCode.OK,
+                        BasicApiResponse(
+                            successful = false,
+                            message = FIELDS_BLANK
                         )
-                    }
-                    is CommentService.ValidationEvent.ErrorCommentTooLong -> {
-                        call.respond(
-                            HttpStatusCode.OK,
-                            BasicApiResponse(
-                                successful = false,
-                                message = COMMENT_TOO_LONG
-                            )
+                    )
+                }
+                is CommentService.ValidationEvent.ErrorCommentTooLong -> {
+                    call.respond(
+                        HttpStatusCode.OK,
+                        BasicApiResponse(
+                            successful = false,
+                            message = COMMENT_TOO_LONG
                         )
-                    }
-                    is CommentService.ValidationEvent.Success -> {
-                        call.respond(
-                            HttpStatusCode.OK,
-                            BasicApiResponse(
-                                successful = true
-                            )
+                    )
+                }
+                is CommentService.ValidationEvent.Success -> {
+                    call.respond(
+                        HttpStatusCode.OK,
+                        BasicApiResponse(
+                            successful = true
                         )
-                    }
+                    )
                 }
             }
         }
@@ -82,7 +76,6 @@ fun Route.getCommentsForPost(
 
 fun Route.deleteComment(
     commentService: CommentService,
-    userService: UserService,
     likeService: LikeService
 ){
     authenticate {
@@ -91,22 +84,22 @@ fun Route.deleteComment(
                 call.respond(HttpStatusCode.BadRequest)
                 return@delete
             }
-            ifEmailBelongsToUser(
-                userId = request.userId,
-                validateEmail = userService::doesEmailBelongToUserId
-            ){
-                val deleted = commentService.deleteComment(request.commentId)
-                if(deleted){
-                    likeService.deleteLikesForParent(request.commentId)
-                    call.respond(
-                        HttpStatusCode.OK,
-                        BasicApiResponse(
-                            successful = true
-                        )
+            val comment = commentService.getCommentById(request.commentId)
+            if(comment?.userId != call.userId){
+                call.respond(HttpStatusCode.Unauthorized)
+                return@delete
+            }
+            val deleted = commentService.deleteComment(request.commentId)
+            if(deleted){
+                likeService.deleteLikesForParent(request.commentId)
+                call.respond(
+                    HttpStatusCode.OK,
+                    BasicApiResponse(
+                        successful = true
                     )
-                }else{
-                    call.respond(HttpStatusCode.NotFound,BasicApiResponse(successful = false))
-                }
+                )
+            }else{
+                call.respond(HttpStatusCode.NotFound,BasicApiResponse(successful = false))
             }
         }
     }
