+++ src/main/kotlin/com/kuluruvineeth/routes/PostRoutes.kt
@@ -5,7 +5,6 @@ import com.kuluruvineeth.data.requests.CreatePostRequest
 import com.kuluruvineeth.data.requests.DeletePostRequest
 import com.kuluruvineeth.data.requests.FollowUpdateRequest
 import com.kuluruvineeth.data.responses.BasicApiResponse
-import com.kuluruvineeth.plugins.email
 import com.kuluruvineeth.repository.post.PostRepository
 import com.kuluruvineeth.service.LikeService
 import com.kuluruvineeth.service.PostService
@@ -23,7 +22,6 @@ import io.ktor.server.routing.*
 
 fun Route.createPost(
     postService: PostService,
-    userService: UserService
 ){
     authenticate {
         post("/api/post/create"){
@@ -31,27 +29,23 @@ fun Route.createPost(
                 call.respond(HttpStatusCode.BadRequest)
                 return@post
             }
-            ifEmailBelongsToUser(
-                userId = request.userId,
-                validateEmail = userService::doesEmailBelongToUserId
-            ){
-                val didUserExist = postService.createPostIfUserExists(request)
-                if(!didUserExist){
-                    call.respond(
-                        HttpStatusCode.OK,
-                        BasicApiResponse(
-                            successful = false,
-                            message = ApiResponseMessages.USER_NOT_FOUND
-                        )
+            val userId = call.userId
+            val didUserExist = postService.createPostIfUserExists(request,userId)
+            if(!didUserExist){
+                call.respond(
+                    HttpStatusCode.OK,
+                    BasicApiResponse(
+                        successful = false,
+                        message = ApiResponseMessages.USER_NOT_FOUND
                     )
-                }else{
-                    call.respond(
-                        HttpStatusCode.OK,
-                        BasicApiResponse(
-                            successful = true,
-                        )
+                )
+            }else{
+                call.respond(
+                    HttpStatusCode.OK,
+                    BasicApiResponse(
+                        successful = true,
                     )
-                }
+                )
             }
 
 
@@ -62,27 +56,17 @@ fun Route.createPost(
 
 fun Route.getPostsForFollows(
     postService: PostService,
-    userService: UserService
 ){
     authenticate {
         get("/api/post/get"){
-            val userId = call.parameters[QueryParams.PARAM_USER_ID] ?: kotlin.run {
-                call.respond(HttpStatusCode.BadRequest)
-                return@get
-            }
             val page = call.parameters[QueryParams.PARAM_PAGE]?.toIntOrNull() ?: 0
             val pageSize = call.parameters[QueryParams.PARAM_PAGE_SIZE]?.toIntOrNull() ?: Constants.DEFAULT_POST_PAGE_SIZE
 
-            ifEmailBelongsToUser(
-                userId = userId,
-                validateEmail = userService::doesEmailBelongToUserId
-            ){
-                val posts = postService.getPostsForFollows(userId,page,pageSize)
-                call.respond(
-                    HttpStatusCode.OK,
-                    posts
-                )
-            }
+            val posts = postService.getPostsForFollows(call.userId,page,pageSize)
+            call.respond(
+                HttpStatusCode.OK,
+                posts
+            )
         }
     }
 }
@@ -90,31 +74,31 @@ fun Route.getPostsForFollows(
 
 fun Route.deletePost(
     postService : PostService,
-    userService: UserService,
     likeService: LikeService
 ){
-    delete("/api/post/delete"){
-        val request = kotlin.runCatching { call.receiveNullable<DeletePostRequest>() }.getOrNull() ?: kotlin.run {
-            call.respond(HttpStatusCode.BadRequest)
-            return@delete
-        }
-        val post = postService.getPost(request.postId)
-        if(post==null){
-            call.respond(
-                HttpStatusCode.NotFound
-            )
-            return@delete
-        }
-        ifEmailBelongsToUser(
-            userId = post.userId,
-            validateEmail = userService::doesEmailBelongToUserId
-        ){
-            postService.deletePost(request.postId)
-            likeService.deleteLikesForParent(request.postId)
-            //TODO: Delete comments from post
-            call.respond(HttpStatusCode.OK)
-        }
+    authenticate {
+        delete("/api/post/delete"){
+            val request = kotlin.runCatching { call.receiveNullable<DeletePostRequest>() }.getOrNull() ?: kotlin.run {
+                call.respond(HttpStatusCode.BadRequest)
+                return@delete
+            }
+            val post = postService.getPost(request.postId)
 
-
+            if(post==null){
+                call.respond(
+                    HttpStatusCode.NotFound
+                )
+                return@delete
+            }
+            if(post.userId == call.userId){
+                postService.deletePost(request.postId)
+                likeService.deleteLikesForParent(request.postId)
+                //TODO: Delete comments from post
+                call.respond(HttpStatusCode.OK)
+            }else{
+                call.respond(HttpStatusCode.Unauthorized)
+            }
+        }
     }
+
 }
